# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "RS Elisabeth CMS API",
        "description": "API documentation for RS Elisabeth CMS endpoints.",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/v1/cms/news": {
            "get": {
                "tags": [
                    "News"
                ],
                "summary": "Get list of news",
                "description": "Returns list of published news with pagination.",
                "operationId": "467e4e02142db57fec31bda53013aefc",
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    }
                },
                "security": [
                    {
                        "HmacAuth": []
                    }
                ]
            }
        },
        "/api/v1/cms/news/{slug}": {
            "get": {
                "tags": [
                    "News"
                ],
                "summary": "Get news detail",
                "description": "Returns a single news data.",
                "operationId": "30ee2d58fbc129112785e92a25afcbd7",
                "parameters": [
                    {
                        "name": "slug",
                        "in": "path",
                        "description": "News slug",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    },
                    "404": {
                        "description": "Resource Not Found"
                    }
                },
                "security": [
                    {
                        "HmacAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "HmacAuth": {
                "type": "apiKey",
                "name": "X-Signature",
                "in": "header"
            }
        }
    },
...
}